update event components
update and merge event components
