+++ src/Event.js
@@ -1,5 +1,46 @@
 import './Event.css';
-const Event = ({ event }) => {
+import Modal from './Modal';
+import Tag from './Tag';
+
+const EventTag = ({ event }) => {
+  return (
+    <div className='event-modal-header'>
+      <Tag name={event.channel_id} color='#d4515d' />
+    </div>
+  );
+};
+const EventContent = ({ event }) => {
+  const start = event.start_date;
+  const end = event.due_date;
+  const options = {
+    weekday: 'short',
+    year: 'numeric',
+    month: 'long',
+    day: 'numeric',
+  };
+  return (
+    <div className='event-modal-content'>
+      <h3 className='event-title'>{event.title}</h3>
+      <div className='event-date'>
+        {start.toLocaleDateString('ko-KR', options)} ~
+      </div>
+      <div className='event-date'>
+        {end.toLocaleDateString('ko-KR', options)}
+      </div>
+      <div>{/*메모*/}</div>
+    </div>
+  );
+};
+export const EventModal = ({ isActive, event }) => {
+  return (
+    <Modal
+      isActive={isActive}
+      header={<EventTag event={event} />}
+      content={<EventContent event={event} />}
+    ></Modal>
+  );
+};
+export const EventList = ({ event }) => {
   const options = {
     weekday: 'short',
     month: 'long',
@@ -20,4 +61,3 @@ const Event = ({ event }) => {
     </div>
   );
 };
-export default Event;
